move some util functions from evaluate.py to utils.py
